+++ src/client/js/index.js
@@ -7,3 +7,4 @@ import '../../common/js/index.js';
 import './blaze.js';
 import './functions.js';
 import './startup.js';
+import './warn_deprecation.js';